Use Jekyll and exclude README, LICENCE, CNAME
Originally I just had Jekyll disabled in order to publish folders that
began with a '.' (specifically .well-known) but that means that other
git artifacts such as the README and LICENCE will be included in the
published site. So instead I'm now using _config.yml to specify files /
folders to include or exclude on the built site.